os/bluestore: bluefs_transaction version 2 for new WAL
New fnode doesn't really work as advertised because there are no
assertions when trying to read with a wrong version with api
`DENC_START...`. Thankfully (or not), bluefs_transaction does have the
good api `DECODE_START`. This way we ensure when downgrading, previous
versions don't try to read this new WAL files.

Signed-off-by: Pere Diaz Bou <pere-altea@hotmail.com>